Introduce rich for text formatting.
`Context` now provides access to a `rich.console.Console` instance of the parser
for pretty text formatting.

`Parser` can be initialized with a `rich.theme.Theme` instance in the
`theme` argument. This will set the theme for the parser's console
instance. clicore provides a default style when a theme isn't passed.

`Context` implements some shortcut methods for quick formatting.